chore: setup type-safe CI with biome
