+++ apps/discord-bot/package.json
@@ -1,14 +1,15 @@
 {
-  "name": "@draft/discord-todo-bot",
-  "version": "0.1.0",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "start": "node index.js",
-    "test": "echo \"No tests\" && exit 0"
-  },
-  "dependencies": {
-    "discord.js": "^14.14.1",
-    "dotenv": "^16.4.5"
-  }
+	"name": "@draft/discord-todo-bot",
+	"version": "0.1.0",
+	"main": "dist/index.js",
+	"type": "module",
+	"scripts": {
+		"build": "tsc -p tsconfig.json",
+		"start": "node dist/index.js",
+		"test": "echo \"No tests\" && exit 0"
+	},
+	"dependencies": {
+		"discord.js": "^14.14.1",
+		"dotenv": "^16.4.5"
+	}
 }
