+++ biome.json
@@ -1,0 +1,6 @@
+{
+	"$schema": "https://biomejs.dev/schemas/1.7.0/schema.json",
+	"files": {
+		"ignore": ["node_modules", "**/dist"]
+	}
+}
